requirements.txt
dataclass_csv
numpy
pandas
# fhirpy
# fhir.resources
geopy
chronyk
requests
mysql.connector
folium
openpyxl
names
postgres==4.0
# mysql-connector
matplotlib==3.8.2
random-address==1.1.1
phone-gen==2.3.15
sqlalchemy-iris
json